partially implemented file manager into pager and started migrating from usize to u64

+++ src/size_constants.rs
@@ -1,5 +1,5 @@
 pub const PAGE_SIZE: u64 = 4096;
-pub const TABLE_MAX_PAGES: usize = 100;
-pub const ROW_SIZE: usize = 307;
-pub const ROWS_PER_PAGE: usize = (PAGE_SIZE as usize) / ROW_SIZE; // Evaluates to 13 rows
-pub const TABLE_MAX_ROWS: usize = ROWS_PER_PAGE * TABLE_MAX_PAGES;
+pub const TABLE_MAX_PAGES: u64 = 100;
+pub const ROW_SIZE: u64 = 307;
+pub const ROWS_PER_PAGE: u64 = PAGE_SIZE / ROW_SIZE; // Evaluates to 13 rows
+pub const TABLE_MAX_ROWS: u64 = ROWS_PER_PAGE * TABLE_MAX_PAGES;
